requirements.txt
openai
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
